Moved the hooks used for the editing feature from the shared folder to Edit Flashcard Set folder
